Vanilla script updated for self captures
self captures and assuming that the data contains ALL comparisons, not just unique ones. the likleihood therefore only considers a single direction. Also, pop_found is replaced by kinship, thus allowing for multiple kinships to be used (so far, U and S are U, and PO/OP is the other)